pyproject.toml
[tool.poetry]
name = "pygraz-website"
version = "0.1.0"
description = "Website of the Python user group Graz"
authors = ["Thomas Aglassinger <roskakori@users.sourceforge.net>"]
license = "CC-BY"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
lektor = "^3.3.10"
python-slugify = "^8.0.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
